portal: update dependencies of all examples + some refactoring
