+++ packages/examples/example-ssr/src/App.jsx
@@ -8,8 +8,8 @@ import PropTypes from 'prop-types'
 import { PrerenderedComponent } from 'react-prerendered-component'
 
 // With this we get 100kb more in bundle size
-import { Button, Input, Icon } from 'dnb-ui-lib/components'
-import { H1, H2, P, Section } from 'dnb-ui-lib/elements'
+import { Button, Input, Icon, Section } from 'dnb-ui-lib/components'
+import { H1, H2, P } from 'dnb-ui-lib/elements'
 import { bell_medium as Bell } from 'dnb-ui-lib/icons'
 
 export default class App extends PureComponent {
@@ -44,7 +44,7 @@ export default class App extends PureComponent {
         live={this.state.restored}
       >
         <div className="dnb-core-style">
-          <Section className="dnb-spacing" useSpacing>
+          <Section className="dnb-spacing" spacing="true">
             <H1>React Components {this.state.count}</H1>
             <P>
               This is not for real world usage. But only to show the
@@ -61,7 +61,7 @@ export default class App extends PureComponent {
               on_click={this.handleClick}
             />
           </Section>
-          <Section className="dnb-spacing" useSpacing>
+          <Section className="dnb-spacing" spacing="true">
             <H2>H2</H2>
             <P>
               Show me an Icon <Icon icon={Bell} size="medium" />
