+++ packages/examples/example-styling/package.json
@@ -13,18 +13,18 @@
   },
   "dependencies": {
     "dnb-ui-lib": "*",
-    "react": "^16.8.3",
-    "react-dom": "^16.8.3"
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6"
   },
   "devDependencies": {
-    "@babel/core": "^7.4.3",
-    "@babel/plugin-proposal-class-properties": "^7.4.0",
+    "@babel/core": "^7.4.5",
+    "@babel/plugin-proposal-class-properties": "^7.4.4",
     "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "@babel/polyfill": "^7.4.3",
-    "@babel/preset-env": "^7.4.3",
+    "@babel/polyfill": "^7.4.4",
+    "@babel/preset-env": "^7.4.5",
     "@babel/preset-react": "^7.0.0",
-    "@emotion/core": "^10.0.7",
-    "@emotion/styled": "^10.0.7",
-    "parcel-bundler": "^1.11.0"
+    "@emotion/core": "^10.0.14",
+    "@emotion/styled": "^10.0.14",
+    "parcel-bundler": "^1.12.3"
   }
 }
