+++ packages/examples/example-vue/babel.config.js
@@ -6,9 +6,7 @@ module.exports = {
         debug: false,
         targets: {
           browsers: ['last 2 versions', 'explorer >= 11']
-        },
-        useBuiltIns: 'usage', // add polyfill proposal usage
-        corejs: 2 // define the version we use in package.json
+        }
       }
     ],
     'vue'
