+++ packages/examples/example-vue/package.json
@@ -12,35 +12,33 @@
     "build": "webpack --mode=production --display-modules"
   },
   "dependencies": {
-    "core-js": "2",
     "dnb-ui-lib": "*",
-    "react": "^16.8.4",
-    "react-dom": "^16.8.4",
-    "vue": "^2.6.9",
-    "vuera": "^0.2.3"
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
+    "vue": "^2.6.10",
+    "vuera": "^0.2.5"
   },
   "devDependencies": {
-    "@babel/core": "^7.4.3",
-    "@babel/plugin-proposal-class-properties": "^7.4.0",
-    "@babel/polyfill": "^7.4.3",
-    "@babel/preset-env": "^7.4.3",
+    "@babel/core": "^7.4.5",
+    "@babel/plugin-proposal-class-properties": "^7.4.4",
+    "@babel/preset-env": "^7.4.5",
     "@babel/preset-react": "^7.0.0",
-    "@vue/component-compiler-utils": "^2.6.0",
+    "@vue/component-compiler-utils": "^3.0.0",
     "babel-preset-vue": "^2.0.2",
-    "babel-loader": "^8.0.5",
-    "css-loader": "^2.1.1",
-    "eslint-plugin-vue": "^5.2.2",
-    "file-loader": "^3.0.1",
-    "node-sass": "^4.11.0",
+    "babel-loader": "^8.0.6",
+    "css-loader": "^3.0.0",
+    "eslint-plugin-vue": "^5.2.3",
+    "file-loader": "^4.0.0",
+    "node-sass": "^4.12.0",
     "prop-types": "^15.7.2",
     "sass-loader": "^7.1.0",
     "style-loader": "^0.23.1",
     "vue-hot-reload-api": "^2.3.3",
     "vue-loader": "^15.7.0",
     "vue-style-loader": "^4.1.2",
-    "vue-template-compiler": "^2.6.9",
-    "webpack": "^4.29.6",
-    "webpack-cli": "^3.3.0",
-    "webpack-dev-server": "^3.2.1"
+    "vue-template-compiler": "^2.6.10",
+    "webpack": "^4.35.2",
+    "webpack-cli": "^3.3.5",
+    "webpack-dev-server": "^3.7.2"
   }
 }
